day divider: move code to a new file
Only code motion, no changes in functionality.
